feature: adiciona react router e refatora componentes
